+++ e2e/tests/contracts.rs
@@ -679,7 +679,7 @@ async fn test_connect_wallet() -> Result<()> {
 
     // confirm that funds have been deducted
     let wallet_balance = wallet.get_asset_balance(&Default::default()).await?;
-    assert!(DEFAULT_COIN_AMOUNT > wallet_balance);
+    assert!(DEFAULT_COIN_AMOUNT as u128 > wallet_balance);
 
     // pay for call with wallet_2
     contract_instance
@@ -694,7 +694,7 @@ async fn test_connect_wallet() -> Result<()> {
     let wallet_balance_second_call = wallet.get_asset_balance(&Default::default()).await?;
     let wallet_2_balance = wallet_2.get_asset_balance(&Default::default()).await?;
     assert_eq!(wallet_balance_second_call, wallet_balance);
-    assert!(DEFAULT_COIN_AMOUNT > wallet_2_balance);
+    assert!(DEFAULT_COIN_AMOUNT as u128 > wallet_2_balance);
 
     Ok(())
 }
@@ -760,7 +760,7 @@ async fn test_output_variable_estimation() -> Result<()> {
 
         for wallet in wallets.iter() {
             let balance = wallet.get_asset_balance(&mint_asset_id).await?;
-            assert_eq!(balance, amount);
+            assert_eq!(balance, amount as u128);
         }
     }
 
@@ -810,7 +810,7 @@ async fn test_output_variable_estimation_multicall() -> Result<()> {
 
     for wallet in wallets.iter() {
         let balance = wallet.get_asset_balance(&mint_asset_id).await?;
-        assert_eq!(balance, 3 * amount);
+        assert_eq!(balance, 3 * amount as u128);
     }
 
     Ok(())
@@ -1131,13 +1131,13 @@ async fn test_add_custom_assets() -> Result<()> {
 
     let balance_asset_1 = wallet_1.get_asset_balance(&asset_id_1).await?;
     let balance_asset_2 = wallet_1.get_asset_balance(&asset_id_2).await?;
-    assert_eq!(balance_asset_1, initial_amount - amount_1);
-    assert_eq!(balance_asset_2, initial_amount - amount_2);
+    assert_eq!(balance_asset_1, (initial_amount - amount_1) as u128);
+    assert_eq!(balance_asset_2, (initial_amount - amount_2) as u128);
 
     let balance_asset_1 = wallet_2.get_asset_balance(&asset_id_1).await?;
     let balance_asset_2 = wallet_2.get_asset_balance(&asset_id_2).await?;
-    assert_eq!(balance_asset_1, initial_amount + amount_1);
-    assert_eq!(balance_asset_2, initial_amount + amount_2);
+    assert_eq!(balance_asset_1, (initial_amount + amount_1) as u128);
+    assert_eq!(balance_asset_2, (initial_amount + amount_2) as u128);
 
     Ok(())
 }
