+++ examples/wallets/src/lib.rs
@@ -314,7 +314,7 @@ mod tests {
         let wallet = launch_provider_and_get_wallet().await?;
         // ANCHOR: get_asset_balance
         let asset_id = AssetId::zeroed();
-        let balance: u64 = wallet.get_asset_balance(&asset_id).await?;
+        let balance: u128 = wallet.get_asset_balance(&asset_id).await?;
         // ANCHOR_END: get_asset_balance
 
         // ANCHOR: get_balances
